настройки лимитера вынесены в отдельный файл
