+++ app.js
@@ -3,8 +3,7 @@ const express = require('express');
 const mongoose = require('mongoose');
 
 mongoose.set('strictQuery', true);
-// const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
+const helmet = require('helmet');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 
@@ -18,14 +17,10 @@ const errorsHandler = require('./errors/errorHandler');
 const auth = require('./middlewares/auth');
 const routes = require('./routes');
 
-const { PORT = 3000, DATABASE } = process.env;
+const { PORT = 3000, DATABASE = 'bitfilmsdb' } = process.env;
 const app = express();
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 100, // Limit each IP to 100 requests per `window` (per 15 minutes)
-  standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-});
+const { limiter } = require('./utils/rate-limiter');
+
 mongoose.connect(`mongodb://127.0.0.1:27017/${DATABASE}`);
 
 // const options = {
@@ -42,13 +37,13 @@ mongoose.connect(`mongodb://127.0.0.1:27017/${DATABASE}`);
 // };
 
 // app.use('*', cors(options));
+app.use(requestLogger);
 app.use(limiter);
 app.use(cookieParser());
 
 app.use(bodyParser.json());
 
-app.use(requestLogger);
-// app.use(helmet());
+app.use(helmet());
 app.use(routes);
 
 app.use(auth);
